requirements.txt
livekit-agents[deepgram,openai,silero]~=1.0
python-dotenv~=1.0
